no secret source
There is no secret source.  The code used for "Firefox automatic install for Linux" is used in "Thunder automatic install for Linux".  The only difference is the file names, paths, and URL for download.  If you were expecting new voodoo magic, I'm sorry to disappoint. But the cool thing, you don't have to install it yourself. It's all complete at the push of a few numbers. lol