patch fix
Thunderbird Daily has no CDN redirect for reasons unknown. Until this is resolved, the user will be instructed to download the tar.bz2 archive and the installer can proceed from there. It's not the greatest solution, and this issue thankfully does not impact the stable or beta build, only the daily/nightly build.